refactor: extract DSparkConfidenceHead into its own header.
